requirements.txt
cadquery-ocp>=7.8.1
pythonocc-core>=7.7.2; platform_system=="Darwin" or platform_system=="Windows"
